addMany should also run replaceStreamObjects over activity
